requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
alt-profanity-check==1.5.1
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
azure-core==1.30.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
build==1.2.1
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cloudpathlib==0.18.1
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
confection==0.1.5
coverage==7.6.0
cryptography==43.0.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
dataclasses-json==0.6.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
detect-secrets==1.5.0
detoxify==0.5.2
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
entmax==1.3
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
Faker==25.9.2
fast-langdetect==0.2.1
fastapi==0.112.0
fasttext-wheel==0.9.2
filelock==3.15.4
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
googlemaps==4.10.0
greenlet==3.0.3
griffe==0.36.9
grpcio==1.65.4
guardrails-ai==0.5.2
guardrails-api-client==0.3.9
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
iso-language-codes==1.1.0
isoduration==20.11.0
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==3.13.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.3
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.96
language_data==1.2.0
lightning-utilities==0.11.6
litellm==1.40.8
llama-cloud==0.0.11
llama-index==0.10.59
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.59
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.32
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
lxml==4.9.4
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.38.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
phonenumbers==8.13.42
pillow==10.4.0
pinecone-client==3.2.2
pip-tools==7.4.1
pluggy==1.5.0
portalocker==2.10.1
posthog==3.5.0
preshed==3.0.9
presidio_analyzer==2.2.355
presidio_anonymizer==2.2.355
protobuf==4.25.4
py-readability-metrics==1.4.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydash==7.0.7
Pygments==2.18.0
PyJWT==2.9.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pysqlite3-binary==0.5.3.post1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.5
rebuff @ git+https://github.com/protectai/rebuff@4d2fe064abf164e7381556d23e48e210080f8afa#subdirectory=python-sdk
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.7.1
robust-downloader==0.0.2
rpds-py==0.19.1
rsa==4.9
rstr==3.2.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
spacy==3.7.5
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
sqlglot==25.8.1
sqlparse==0.5.1
sqlvalidator==0.0.20
srsly==2.4.8
starlette==0.37.2
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
thefuzz==0.22.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.5.2
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchmetrics==0.10.3
tqdm==4.66.4
transformers==4.43.3
triton==3.0.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unbabel-comet==2.2.2
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
